backend/requirements.txt
Flask==0.11.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
marshmallow==3.14.1
user-agents==2.2.0
python-dotenv==0.19.0
